Rewrite slightly confusing self-link attribute code
The idea is as follows: This code creates self-links that point
back to the special page, but with one of the URL parameters being
modified. It's used two times, to change the limit=… and the
redirects=… flag.

Parameters that don't change need to be in the link as well. Except
for redirects=1 which is the default.

Change-Id: I330ac7596faa11876a289a524afb6405d3db94ca